+++ modules/audioset/src/migrate.js
@@ -4,7 +4,9 @@ const HSV = /^hsv\((.*)\)$/;
 function rgb(color) {
   const parsed = HSV.exec(color)[1]
     .split(",")
-    .map(i => parseInt(i, 10));
+    .map(function(i) {
+      return parseInt(i, 10);
+    });
   const c = hsvToRgb.apply(null, parsed);
   return "rgb(" + c[0] + "," + c[1] + "," + c[2] + ")";
 }
@@ -19,7 +21,7 @@ function resourceUrl(name, resource) {
 
 export default function migrate(set) {
   console.log("migrate", set.id);
-  Object.keys(set.clips).forEach(name => {
+  Object.keys(set.clips).forEach(function(name) {
     const clip = set.clips[name];
     clip.id = name;
     clip.display.color = rgb(clip.display.color);